views: Add support for an "Attachment" message
Attachment messages currently come from actors and can contain
both a path and a description. The description tells the user
what the file is about and the path is used to compute both the
file name and the icon.

Upon clicking on the icon, the default handler for that file type
is opened.

https://phabricator.endlessm.com/T14182