main: Add QueueableBox class
This class is a simple extention to GtkBox that queues up its
children to be displayed later. Calling pushWidget when there
are no widgets in the queue will display that widget immediately
and future calls to pushWidget will buffer until showNext is called
whereby the next widget in the queue will be shown.

We'll use this to implement the pending message bubbles and ensure
that only one pending message bubble is displayed at a time whilst
the other messages are buffered up.