main: Make _addItem take a timeout and visibility handler
We now integrate with QueueableBox to push any newly created
widgets here on to a queue. When the widget becomes visible
in the sense that the pending handler has completed, we'll start
the countdown for user interaction here. We also call the visibility
handler so that callers can determine whether to, eg, send
notifications or set the unread flag.

https://phabricator.endlessm.com/T14866