main: Show user's input in the chatbox history on submit
As part of this change, we hide the user input bubble and show
the response in the chat history. In order to ensure that the
user input bubble is displayed again at the right time, we add
a function to QueueableBox which is called when the user input
bubble should become available.