views: Display no padding around thumbnails and a border around icons
We set different style classes here depending on the content. We also
change the layout of the content box from horizontal to vertical if
we are displaying an image.

In the CSS, we don't have any padding for thumbnail attachments. The
padding looks out of place and not having padding is analagous to
the behaviour of many other chat apps which don't have padding
around 'media-preview' bubbles.