views: Save thumbnail file directly instead of using DesktopThumbnailFactory
Using DesktopThumbnailFactory doesn't guarantee that the right-sized
thumbnail will be saved for our image. We always want a large
thumbnail and we always want it to be saved as a large thumbnail.

Doing so this way also allows us to keep track of the thumbnail path
as soon as the thumbnail is saved, which can then be passed on to
the CSS (instead of passing null in some cases)

https://phabricator.endlessm.com/T15359