views: Set can-focus to false on message views
It doesn't make any sense to give these the input focus and
doing so causes the parent scroll view to scroll to that message
view when the conversation is focussed. In some cases, that can
result in the conversation window being scrolled to the top
because the top message received the input focus.

https://phabricator.endlessm.com/T15567